+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added "Celestial Body" controls to the Directional Light
 - Added new parameters to the Physically Based Sky
 - Added Reflections to the DXR Wizard
+- Added the possibility to have ray traced colored and semi-transparent shadows on directional lights.
 
 ### Fixed
 - Sorting, undo, labels, layout in the Lighting Explorer.